Performance History
Basic Usage:
    Set up a database containing records of combinations of machine properties and gufi configurations
    Run GUFI and redirect the debug output into the appropriate raw data database
    plot statistics using the data stored in the raw data database

    See example.sh

Run pylint on python files to check style
    cleaned up scripts
moved hashes dictionary out of trace_anonymizer
