+++ .github/workflows/check.yml
@@ -157,6 +157,33 @@ jobs:
       working-directory: ${{ github.workspace }}/build
       run:  ctest --output-on-failure
 
+  Pylint:
+    runs-on: ubuntu-20.04
+    steps:
+    - uses: actions/checkout@v3
+
+    - uses: actions/cache@v3
+      with:
+        path: ${{ env.DEP_INSTALL_PREFIX }}
+        key:  ${{ runner.os }}-pylint
+
+    - name: Ubuntu 20.04 Prerequisites
+      run: sudo contrib/CI/ubuntu.sh
+
+    - name: Install Extra Python Packages
+      run: |
+        sudo apt -y install python3-pip python3-matplotlib
+        sudo pip3 install pylint
+
+    - name: Configure CMake
+      run:  cmake -B ${{ github.workspace }}/build ${{ env.COMMON_CONFIG }} -DTEST_WORKING_DIRECTORY=/tmp
+
+    - name: Build
+      run:  cmake --build ${{ github.workspace }}/build -j
+
+    - name: Run Python3 Pylint
+      run: contrib/CI/pylint.sh ${{ github.workspace }}/build
+
   DebugPrints:
     runs-on: ubuntu-20.04
     steps:
@@ -215,4 +242,4 @@ jobs:
       run:  sudo apt -y install shellcheck
 
     - name: Test
-      run:  shellcheck -s bash -e SC1090,SC1091 $(find \( -name "*.sh" -o -name "*.sh.in" \) -not -path "*/contrib/debug/*")
+      run:  shellcheck -s bash -e SC1090,SC1091,SC2086 $(find \( -name "*.sh" -o -name "*.sh.in" \) -not -path "*/contrib/debug/*")
