+++ contrib/performance/configs/db_test.ini
@@ -1,0 +1,32 @@
+[raw_data]
+commits = commit-1, commit-2, commit-3, commit-4
+columns = Total Thread Time (not including main), Real time (main)
+
+[output]
+path = ./db_test.png
+graph_title = Basic Graph Example
+graph_dimensions = 12, 6
+
+[line]
+colors = green
+types = solid
+markers = o
+
+[axes]
+hash_len = 0
+y_label = Time (seconds)
+y_min = 0
+y_max = 30
+
+[annotations]
+show = True
+precision = 2
+x_offset = 5
+y_offset = 5
+text_colors = orange
+
+[error_bar]
+show = True
+min_max = True
+precision = 2
+text_colors = green
