+++ contrib/performance/configs/db_test_complex.ini
@@ -1,0 +1,32 @@
+[raw_data]
+commits = commit-1, commit-2, commit-3, commit-4
+columns = Total Thread Time (not including main), Real time (main), set up globals
+
+[output]
+path = ./db_test_complex.png
+graph_title = Basic Graph Example
+dimensions = 12, 6
+
+[line]
+colors = green, red
+types = solid, dashed
+markers = o
+
+[axes]
+hash_len = 0
+y_label = Time (seconds)
+y_min = 0
+y_max = 30
+
+[annotations]
+show = True
+precision = 3
+x_offset = 5
+y_offset = 5
+text_colors = blue, green
+
+[error_bar]
+show = False
+min_max = False
+precision = 3
+text_colors = black
