+++ contrib/performance/performance_pkg/common.py.in
@@ -61,11 +61,57 @@
 
 
 
+import os
+import subprocess
 import sys
 
-import common
+from hashes import Hashes # not part of performance_pkg
 
-# do not call this file directly - script will set up environment
-if __name__ == '__main__':
-    gufi_getfattr = common.import_tool('gufi_getfattr')
-    sys.exit(gufi_getfattr.run(sys.argv, common.CONFIG_PATH))
+if (sys.version_info.major >= 3) and (sys.version_info.minor >= 3):
+    DEVNULL = subprocess.DEVNULL # pylint: disable=no-member
+else:
+    DEVNULL = None
+
+TYPE_TO_SQLITE = {
+    bool : 'BOOLEAN',
+    float: 'FLOAT',
+    int  : 'INTEGER',
+    str  : 'TEXT',
+    None : 'INTEGER PRIMARY KEY',
+}
+
+def hash_config(alg, data):
+    return Hashes[alg](data.encode()).hexdigest()
+
+def check_exists(path):
+    if not os.path.exists(path):
+        print("'{0}' does not exist!".format(path)) # pylint: disable=superfluous-parens
+        sys.exit(1)
+
+    if not os.path.isfile(path):
+        print("'{0}' is not a file!".format(path)) # pylint: disable=superfluous-parens
+        sys.exit(2)
+
+def check_not_exists(path):
+    if os.path.exists(path):
+        print("'{0}' already exists!".format(path)) # pylint: disable=superfluous-parens
+        sys.exit(1)
+
+def run_get_stdout(command):
+    # pylint: disable=consider-using-with
+    process = subprocess.Popen(command,
+                               stdout=subprocess.PIPE,
+                               stderr=DEVNULL)
+    out, _ = process.communicate()
+    return out.decode('ascii')
+
+def get_current_commit():
+    commit_hash = run_get_stdout(['git', 'rev-parse', 'HEAD'])[:-1]
+
+    # pylint: disable=consider-using-with
+    diff = subprocess.Popen(['git', 'diff-index', '--quiet', 'HEAD', '--'])
+    diff.communicate() # no output
+    if diff.returncode != 0:
+        commit_hash += '~dirty'
+
+    return commit_hash
