+++ contrib/performance/raw_data_db.py.in
@@ -61,11 +61,41 @@
 
 
 
+import argparse
+import sqlite3
 import sys
 
-import common
+from performance_pkg import common
+from performance_pkg.gufi_query import cumulative_times
+from performance_pkg.hashdb import gufi, utils as hashdb
 
-# do not call this file directly - script will set up environment
+def parse_args(argv):
+    parser = argparse.ArgumentParser()
+    parser.add_argument('database',
+                        type=str,
+                        help='Hash database of configurations (must already exist)')
+    parser.add_argument('raw_data_hash',
+                        help='Hash of a single configuration')
+    parser.add_argument('raw_data_db',
+                        type=str,
+                        help='Raw data database (must already exist)')
+    return parser.parse_args(argv)
+
+def run(argv):
+    args = parse_args(argv)
+
+    gufi_cmd, debug_name = hashdb.get_config(args.database, args.raw_data_hash)
+    print('{0} was run with {1}, debug name {2}'.format(args.raw_data_hash, gufi_cmd, debug_name)) # pylint: disable=superfluous-parens
+
+    common.check_not_exists(args.raw_data_db)
+    try:
+        con = sqlite3.connect(args.raw_data_db)
+        if gufi_cmd == gufi.GUFI_QUERY:
+            if debug_name == gufi.CUMULATIVE_TIMES:
+                cumulative_times.create_tables(con)
+        con.commit()
+    finally:
+        con.close()
+
 if __name__ == '__main__':
-    gufi_getfattr = common.import_tool('gufi_getfattr')
-    sys.exit(gufi_getfattr.run(sys.argv, common.CONFIG_PATH))
+    run(sys.argv[1:])
