+++ scripts/gufi_ls
@@ -105,6 +105,7 @@ def select_time(style, name):
     try:
         column = 'strftime("{0}", {{0}})'.format(TIME_STYLES[style])
     except:
+        # pylint: disable=raise-missing-from
         raise ValueError('gufi_ls: invalid argument \'{0}\' for \'time style\''.format(style))
 
     return [column.format(name)]
@@ -164,13 +165,15 @@ def build_where(args, table, name=None):
     if name:
         where = ['{0}name REGEXP "^{1}$"'.format(table, name)]
     else:
+        # pylint: disable=anomalous-backslash-in-string
+
         # -a: if show all, override -A
         if args.all:
             pass
         else:
             # -A: ignore . and ..
             if args.almost_all:
-                where += ['{0}name <> "." AND {1}name <> ".."'.format(table)]
+                where += ['{0}name <> "." AND {0}name <> ".."'.format(table)]
 
             # everything that's not hidden
             else:
@@ -202,11 +205,14 @@ def build_order_by(args):
 
 # argv[0] should be the command name
 def run(argv, config_path):
+    # pylint: disable=too-many-statements,too-many-locals,invalid-name
+
     # find and parse the configuration file first
     config = gufi_config.Server(config_path)
 
-    class FullTimeAction(argparse.Action):
+    class FullTimeAction(argparse.Action): # pylint: disable=too-few-public-methods
         def __init__(self, option_strings, dest, nargs=None, **kwargs):
+            # pylint: disable=super-with-arguments
             super(FullTimeAction, self).__init__(option_strings, dest, 0, **kwargs)
 
         def __call__(self, parser, namespace, values, option_string=None):
@@ -340,7 +346,7 @@ def run(argv, config_path):
         create_table_cols = ', '.join(
             [
                 '{0} {1}'.format(column, type)
-                for column, type in zip(columns,
+                for column, type in zip(columns, # pylint: disable=redefined-builtin
                                         ['TEXT', 'TEXT', 'TEXT', 'INT64', 'INT64',
                                          'INT64', 'INT64', 'INT64', 'INT64', 'INT64', 'INT64',
                                          'INT64', 'INT64', 'INT64', 'TEXT',
@@ -409,8 +415,8 @@ def run(argv, config_path):
         if args.skip:
             query_cmd += ['-k', args.skip]
 
-        query = subprocess.Popen(query_cmd + [fullpath])
-        query.communicate() # block until query finishes
+        query = subprocess.Popen(query_cmd + [fullpath]) # pylint: disable=consider-using-with
+        query.communicate()                              # block until query finishes
 
         if query.returncode != 0:
             rc = 2
