+++ test/regression/gufi_stats.py.in
@@ -68,4 +68,4 @@ import common
 # do not call this file directly - script will set up environment
 if __name__ == '__main__':
     gufi_stats = common.import_tool('gufi_stats')
-    sys.exit(gufi_stats.run(sys.argv, common.config_path))
+    sys.exit(gufi_stats.run(sys.argv, common.CONFIG_PATH))
